completed hw7
