+++ public/index.html
@@ -10,7 +10,7 @@
       href="https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Nunito:wght@400;500;600;700;800;900&display=swap"
       rel="stylesheet"
     />
-  </head>
+  <script defer src="bundle.js"></script></head>
   <script src="public/bundle.js"></script>
 
   <body>
