+++ src/AddTaskModal/add-task-modal.css
@@ -1,0 +1,139 @@
+.modal {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgba(0, 0, 0, 0.8);
+  }
+  
+  .modal__content {
+    position: relative;
+    margin: 5rem auto;
+    padding: 24px 32px;
+    background-color: #fefefe;
+    border-radius: 24px;
+    border: 1px solid #888;
+    min-height: 350px;
+    max-height: 500px;
+    width: 600px;
+    height: auto;
+  }
+  
+  .modal__content__header {
+    text-align: center;
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 36px;
+  }
+  
+  .modal__content__task-input {
+    width: 85%;
+    /* margin: 38px 27px; */
+    /* margin: 38px 27px 0 27px; */
+    padding: 16px 21px;
+    border-radius: 12px;
+    border: 1px solid #d2d2d2;
+    font-size: 16px;
+    font-weight: 500;
+    color: #838383;
+    background: #f5f5f5;
+    position: absolute;
+  }
+  .modal__content__calendar {
+    border: 2px solid rgba(60, 134, 244, 0.6);
+    padding: 12px 16px 6px 16px;
+    position: absolute;
+    margin-top: 80px;
+    right: 32px;
+    border-radius: 12px;
+    color: #1d1d1d;
+    font-size: 16px;
+    font-weight: 500;
+  }
+  
+  .modal__content__tag-containers {
+    display: flex;
+    flex-wrap: wrap;
+    position: absolute;
+    margin-top: 90px;
+    font-size: 12px;
+    font-weight: 500;
+  }
+  .tag {
+    cursor: pointer;
+    padding: 2px 8px;
+    margin-right: 12px;
+    border-radius: 6px;
+  }
+  .tag:last-child {
+    margin-right: 0;
+  }
+  
+  .health-tag {
+    color: #0053cf;
+    background-color: rgb(60, 134, 244, 0.31);
+  }
+  
+  .health-selected {
+    border: 1px solid #0053cf;
+  }
+  
+  .work-tag {
+    color: #9747ff;
+    background-color: #e8d7ff;
+  }
+  
+  .work-selected {
+    border: 1px solid #9747ff;
+  }
+  
+  .home-tag {
+    color: #639462;
+    background-color: #e2f6e2;
+  }
+  
+  .home-selected {
+    border: 1px solid #639462;
+  }
+  
+  .other-tag {
+    color: #ea8c00;
+    background-color: #ffecc7;
+  }
+  
+  .other-selected {
+    border: 1px solid #ea8c00;
+  }
+  
+  .btn {
+    text-align: center;
+    font-size: 16px;
+    font-weight: 800;
+    line-height: 20px;
+    padding: 16px 0;
+    border-radius: 12px;
+    position: absolute;
+    bottom: 22px;
+    border: none;
+    cursor: pointer;
+  }
+  
+  .btn--cancel {
+    right: calc(50% + 5px);
+    left: 0;
+    background-color: #fefefe;
+    color: #0053cf;
+    margin-left: 27px;
+  }
+  
+  .btn--add-task {
+    left: calc(50% + 5px);
+    right: 0;
+    color: #ffffff;
+    background-color: #d3d3d3;
+    margin-right: 27px;
+  }
