+++ src/All-tasks/all-tasks.css
@@ -1,0 +1,73 @@
+.checkbox {
+  width: 21px;
+  height: 21px;
+  cursor: pointer;
+  border: 2px solid #d3d3d3;
+  border-radius: 4px;
+}
+.checkbox--checked {
+  background-image: url("/src/assets/checked-checkbox.svg");
+  background-repeat: no-repeat;
+  object-fit: cover;
+  background-color: #d3d3d3;
+  background-position: center;
+}
+
+.task__list__item-info__title {
+  font-size: 18px;
+  font-weight: 400;
+  color: #1d1d1d;
+}
+
+.task__list__item-info__date {
+  color: #838383;
+  font-weight: 500;
+  font-size: 12px;
+}
+
+.task__list__item__delete-svg {
+  margin-left: auto;
+  margin-bottom: 30px;
+  cursor: pointer;
+}
+
+#all-tasks {
+  width: 40%;
+  /* margin-bottom: 32px; */
+}
+.tasks__header {
+  font-size: 24px;
+  font-weight: 700;
+  font-family: Nunito;
+  line-height: 32px;
+  padding-bottom: 16px;
+  margin-top: 42px;
+}
+
+.task__list {
+  list-style-type: none;
+}
+
+.task__list__item-info {
+  display: flex;
+  flex-direction: column;
+}
+
+.task__list__item {
+  margin-bottom: 28px;
+}
+
+.task__list__item__delete-svg{
+  background-repeat: no-repeat;
+  object-fit: cover;
+  background-image: url("/src/assets/bin.svg");
+  
+}
+
+.task__list__item__delete {
+  width: 21px;
+  height: 21px;
+  cursor: pointer;
+  border-radius: 4px;
+}
+
