+++ src/Completed-tasks/completed-tasks.css
@@ -1,0 +1,48 @@
+.checkbox {
+  width: 21px;
+  height: 21px;
+  cursor: pointer;
+  border: 2px solid #d3d3d3;
+  border-radius: 4px;
+}
+.checkbox--checked {
+  background-image: url("/src/assets/checked-checkbox.svg");
+  background-repeat: no-repeat;
+  object-fit: cover;
+  background-color: #d3d3d3;
+  background-position: center;
+}
+
+#completed-tasks {
+  width: 40%;
+}
+.completed__header {
+  font-size: 24px;
+  font-weight: 700;
+  font-family: Nunito;
+  line-height: 32px;
+  padding-bottom: 16px;
+}
+.completed__list {
+  list-style-type: none;
+  color: #838383;
+}
+
+.completed__list__item {
+  margin-bottom: 28px;
+}
+.completed__list__item-info {
+  display: flex;
+  flex-direction: column;
+}
+.completed__list__item-info__title {
+  font-size: 18px;
+  font-weight: 400;
+  color: #838383;
+}
+
+.completed__list__item-info__date {
+  color: #838383;
+  font-weight: 500;
+  font-size: 12px;
+}
