+++ src/GM-modal/gm-modal.css
@@ -1,0 +1,57 @@
+.gm-modal {
+  display: none;
+  flex-direction: column;
+  justify-content: center;
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: scroll;
+  /* overflow-y: auto; */
+  background-color: rgba(0, 0, 0, 0.9);
+}
+.gm-modal__content {
+  position: relative;
+  margin: 5rem auto;
+  padding: 24px 32px;
+  background-color: #fefefe;
+  border-radius: 24px;
+  border: 1px solid #888;
+  min-height: 350px;
+  max-height: 500px;
+  width: 600px;
+  /* height: auto; */
+}
+.task-container {
+  height: 150px;
+  overflow: scroll;
+  cursor: pointer;
+}
+
+.gm-modal__button {
+  position: absolute;
+  bottom: 22px;
+  color: #ffffff;
+  left: 32px;
+  right: 32px;
+  background-color: #3c86f4;
+  /* width: 100%; */
+  margin-top: 50px;
+  text-align: center;
+  font-size: 16px;
+  font-weight: 800;
+  line-height: 20px;
+  padding: 16px 0;
+  border-radius: 12px;
+  bottom: 22px;
+  border: none;
+  cursor: pointer;
+}
+.gm-modal__content__header {
+    text-align: center;
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 36px;
+  }
