+++ src/main.css
@@ -1,0 +1,83 @@
+
+body {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: "Nunito";
+}
+
+.container {
+  padding: 0 4rem;
+  margin: 2rem;
+}
+
+.container__header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.container__header--left {
+  font-size: 32px;
+  font-weight: 700;
+  line-height: 36px;
+}
+
+.container__header--right {
+  display: flex;
+  align-items: center;
+  padding: 10px 22px;
+}
+.container__header--right__img {
+  margin-right: 23px;
+}
+.container__header--right__temp {
+  margin-right: 80px;
+  font-weight: 800;
+  font-size: 18px;
+}
+
+.container__header--right__city {
+  color: #838383;
+  font-weight: 500;
+  font-size: 18px;
+}
+.flex {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.container__search {
+  width: 100%;
+  padding: 16px 21px;
+  margin-right: 40px;
+  border-radius: 12px;
+  border: 1px solid #d2d2d2;
+  font-size: 16px;
+  font-weight: 500;
+  color: #838383;
+  background: #f5f5f5;
+}
+
+.container__new-task {
+  cursor: pointer;
+  background-color: rgba(60, 134, 244, 0.15);
+  padding: 16px 32px;
+  border: none;
+  font-size: 16px;
+  font-weight: 800;
+  color: #0053cf;
+  height: 52px;
+  white-space: nowrap;
+  border-radius: 12px;
+}
+
+.checkbox {
+  margin-right: 20px;
+  width: 21px;
+  height: 21px;
+  color: #d3d3d3; /* default color */
+}
+
+
