+++ src/todo.ts
@@ -1,0 +1,7 @@
+export interface TodoItem {
+    id: number,
+    title: string;
+    isCompleted: boolean;
+    tag: string;
+    date: string; 
+}
